pyproject.toml
[project]
name = "nonebot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cachetools>=6.2.4",
    "distro>=1.9.0",
    "google-genai>=1.56.0",
    "httpx[socks]>=0.28.1",
    "langgraph>=1.0.5",
    "nonebot-adapter-console>=0.9.0",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-alconna>=0.60.3",
    "nonebot2[fastapi]>=2.4.4",
]
